refactored readonly smoketest to use new cognito task
