+++ cypress-smoketests-readonly.config.ts
@@ -1,9 +1,22 @@
 import { defineConfig } from 'cypress';
+import { getUserToken as getUserTokenLocal } from './cypress/cypress-smoketests/support/pages/local-login';
+import { getUserTokenWithRetry } from './cypress/support/cognito-login';
+
+const { CYPRESS_SMOKETESTS_LOCAL } = process.env;
 
 // eslint-disable-next-line import/no-default-export
 export default defineConfig({
-  defaultCommandTimeout: 20000,
+  defaultCommandTimeout: 60000,
   e2e: {
+    setupNodeEvents(on) {
+      on('task', {
+        getUserToken({ email, password }) {
+          return CYPRESS_SMOKETESTS_LOCAL
+            ? getUserTokenLocal(email)
+            : getUserTokenWithRetry(email, password);
+        },
+      });
+    },
     specPattern: 'cypress/cypress-readonly/integration/*.cy.ts',
     supportFile: 'cypress/cypress-readonly/support/index.ts',
     testIsolation: false,
@@ -13,7 +26,7 @@ export default defineConfig({
   reporterOptions: {
     toConsole: true,
   },
-  requestTimeout: 20000,
+  requestTimeout: 60000,
   screenshotsFolder: 'cypress/cypress-readonly/screenshots',
   video: true,
   videoCompression: 10,
