+++ cypress/cypress-readonly/integration/messages.cy.ts
@@ -1,3 +1,4 @@
+import { AuthenticationResult } from '../../support/login-types';
 import { getEnvironmentSpecificFunctions } from '../support/pages/environment-specific-factory';
 import { isValidRequest } from '../support/helpers';
 
@@ -6,28 +7,25 @@ const EFCMS_DOMAIN = Cypress.env('EFCMS_DOMAIN');
 const DEPLOYING_COLOR = Cypress.env('DEPLOYING_COLOR');
 
 describe('Messages UI Smoketests', () => {
-  let token;
+  const { login } = getEnvironmentSpecificFunctions();
 
-  const { getUserToken, login } = getEnvironmentSpecificFunctions();
+  before(() => {
+    cy.intercept({
+      hostname: `api-${DEPLOYING_COLOR}.${EFCMS_DOMAIN}`,
+      method: 'GET',
+      url: '/messages/inbox/*',
+    }).as('getMyMessages');
 
-  before(async () => {
-    let result = await getUserToken(
-      'testAdmissionsClerk@example.com',
-      DEFAULT_ACCOUNT_PASS,
-    );
-    token = result.AuthenticationResult.IdToken;
+    cy.task<AuthenticationResult>('getUserToken', {
+      email: 'testAdmissionsClerk@example.com',
+      password: DEFAULT_ACCOUNT_PASS,
+    }).then(result => {
+      login(result.AuthenticationResult.IdToken);
+    });
   });
 
   describe('login and view the user messages', () => {
     it("should fetch the user's messages after log in", () => {
-      cy.intercept({
-        hostname: `api-${DEPLOYING_COLOR}.${EFCMS_DOMAIN}`,
-        method: 'GET',
-        url: '/messages/inbox/*',
-      }).as('getMyMessages');
-
-      login(token);
-
       cy.wait('@getMyMessages').then(isValidRequest);
     });
