+++ cypress/cypress-readonly/integration/trial-sessions.cy.ts
@@ -1,3 +1,4 @@
+import { AuthenticationResult } from '../../support/login-types';
 import { getEnvironmentSpecificFunctions } from '../support/pages/environment-specific-factory';
 import { isValidRequest } from '../support/helpers';
 
@@ -6,22 +7,19 @@ const EFCMS_DOMAIN = Cypress.env('EFCMS_DOMAIN');
 const DEPLOYING_COLOR = Cypress.env('DEPLOYING_COLOR');
 
 describe('Trial Sessions UI Smoketests', () => {
-  let token;
+  const { login } = getEnvironmentSpecificFunctions();
 
-  const { getUserToken, login } = getEnvironmentSpecificFunctions();
-
-  before(async () => {
-    let result = await getUserToken(
-      'testAdmissionsClerk@example.com',
-      DEFAULT_ACCOUNT_PASS,
-    );
-    token = result.AuthenticationResult.IdToken;
+  before(() => {
+    cy.task<AuthenticationResult>('getUserToken', {
+      email: 'testAdmissionsClerk@example.com',
+      password: DEFAULT_ACCOUNT_PASS,
+    }).then(result => {
+      login(result.AuthenticationResult.IdToken);
+    });
   });
 
   describe('login and view the trial sessions page', () => {
     it('should fetch the open trial sessions upon navigation', () => {
-      login(token);
-
       cy.intercept({
         hostname: `api-${DEPLOYING_COLOR}.${EFCMS_DOMAIN}`,
         method: 'GET',
