+++ cypress/cypress-readonly/support/pages/environment-specific-factory.ts
@@ -1,13 +1,7 @@
-import {
-  getRestApi as getRestApiDeployed,
-  getUserToken as getUserTokenDeployed,
-  login as loginDeployed,
-} from './login';
+import { login as loginDeployed } from './login';
 
 export const getEnvironmentSpecificFunctions = () => {
   return {
-    getRestApi: getRestApiDeployed,
-    getUserToken: getUserTokenDeployed,
     login: loginDeployed,
   };
 };
