+++ cypress/cypress-smoketests/integration/case-creation.cy.ts
@@ -1,3 +1,4 @@
+import { AuthenticationResult } from '../../support/login-types';
 import {
   completeWizardStep1,
   completeWizardStep2,
@@ -14,8 +15,6 @@ import {
   hasIrsNotice,
   submitPetition,
 } from '../support/pages/create-electronic-petition';
-
-import { AuthenticationResult } from '../support/pages/local-login';
 import { faker } from '@faker-js/faker';
 import { fillInCreateCaseFromPaperForm } from '../../cypress-integration/support/pages/create-paper-petition';
 import { getEnvironmentSpecificFunctions } from '../support/pages/environment-specific-factory';
