+++ cypress/cypress-smoketests/integration/change-address.cy.ts
@@ -1,6 +1,5 @@
-import { AuthenticationResult } from '../support/pages/local-login';
+import { AuthenticationResult } from '../../support/login-types';
 import { getEnvironmentSpecificFunctions } from '../support/pages/environment-specific-factory';
-
 import {
   goToEditContactInformation,
   goToMyAccount,
