+++ cypress/cypress-smoketests/support/pages/local-login.ts
@@ -1,9 +1,3 @@
-export type AuthenticationResult = {
-  AuthenticationResult: {
-    IdToken: string;
-  };
-};
-
 export const getUserToken = async username => {
   if (username === 'migrator@example.com') {
     username =
