+++ cypress/support/login-types.ts
@@ -1,0 +1,5 @@
+export type AuthenticationResult = {
+  AuthenticationResult: {
+    IdToken: string;
+  };
+};
